requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow==5.3.0
psycopg2==2.7.4 --no-binary psycopg2

# Django
django==2.2.7

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.9.0
django-filter==2.0.0
djangorestframework-filters==0.11.1

# JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==18.3.0

# Static files
whitenoise==4.1.2

# Celery
redis==3.2.0
django-redis==4.10.0
celery==4.2.1
flower==0.9.2

# AdminCU
cairocffi==0.8.0
CairoSVG==2.1.3
certifi==2018.1.18
cffi==1.11.5
django-cors-headers==3.0.2
cssselect2==0.2.1
django-afip==7.0.0
django-mercadopago==5.1.1
django-templated-email==2.2.0
html5lib==1.0.1
lxml==4.2.1
mercadopago==0.3.4
numpy==1.17.4
pandas==0.23.4
Pyphen==0.9.4
python-barcode==0.8.1
setuptools-git==1.2
setuptools-scm==2.0.0
tablib==0.12.1
tinycss2==0.6.1
tornado==5.1.1
WeasyPrint==0.42.2
xlrd==1.1.0
xlwt==1.3.0
